Move metricslogs component from comp/forwarder to comp/core
comp/forwarder is for components that ship data over the network to
Datadog intake. metricslogs never leaves the host, so comp/core is
the right home for it, alongside other core, non-network-shipping
utilities.

No behavior change: only import paths move.
