+++ cmd/agent/subcommands/run/BUILD.bazel
@@ -71,6 +71,7 @@ go_library(
         "//comp/core/ipc/fx",
         "//comp/core/log/def",
         "//comp/core/lsof/fx",
+        "//comp/core/metricslogs/fx",
         "//comp/core/pid/def",
         "//comp/core/pid/impl",
         "//comp/core/profiler/fx",
@@ -109,7 +110,6 @@ go_library(
         "//comp/forwarder/eventplatform/fx",
         "//comp/forwarder/eventplatform/impl",
         "//comp/forwarder/eventplatformreceiver/impl",
-        "//comp/forwarder/metricslogs/fx",
         "//comp/forwarder/orchestrator/def",
         "//comp/forwarder/orchestrator/fx",
         "//comp/haagent/fx",
