+++ comp/core/metricslogs/def/BUILD.bazel
@@ -3,6 +3,6 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "def",
     srcs = ["component.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/forwarder/metricslogs/def",
+    importpath = "github.com/DataDog/datadog-agent/comp/core/metricslogs/def",
     visibility = ["//visibility:public"],
 )
