+++ comp/core/metricslogs/def/component.go
