+++ comp/core/metricslogs/fx/BUILD.bazel
@@ -3,10 +3,10 @@ load("@rules_go//go:def.bzl", "go_library")
 go_library(
     name = "fx",
     srcs = ["fx.go"],
-    importpath = "github.com/DataDog/datadog-agent/comp/forwarder/metricslogs/fx",
+    importpath = "github.com/DataDog/datadog-agent/comp/core/metricslogs/fx",
     visibility = ["//visibility:public"],
     deps = [
-        "//comp/forwarder/metricslogs/impl",
+        "//comp/core/metricslogs/impl",
         "//pkg/util/fxutil",
     ],
 )
